+++ src/commonMain/kotlin/kotools/types/text/NotBlankString.kt
@@ -46,9 +46,31 @@ public value class NotBlankString private constructor(
     override infix fun compareTo(other: NotBlankString): Int =
         "$this".compareTo("$other")
 
+    /** Concatenates this string with the [other] one. */
+    @SinceKotoolsTypes("4.2")
+    public operator fun plus(other: String): NotBlankString = "$value$other"
+        .toNotBlankString()
+        .getOrThrow()
+
+    /** Concatenates this string with the [other] one. */
+    @SinceKotoolsTypes("4.2")
+    public operator fun plus(other: NotBlankString): NotBlankString =
+        plus("$other")
+
+    /** Concatenates this string with the [other] character. */
+    @SinceKotoolsTypes("4.2")
+    public operator fun plus(other: Char): NotBlankString = plus("$other")
+
     /** Returns this string as a [String]. */
     override fun toString(): String = value
 }
+
+/** Concatenates this character with the [other] string. */
+@SinceKotoolsTypes("4.2")
+public operator fun Char.plus(other: NotBlankString): NotBlankString =
+    plus("$other")
+        .toNotBlankString()
+        .getOrThrow()
 
 /**
  * Returns this string as an encapsulated [NotBlankString], or returns an