+++ src/commonMain/kotlin/kotools/types/number/StrictlyNegativeInt.kt
@@ -4,6 +4,10 @@ import kotlinx.serialization.Serializable
 import kotlinx.serialization.SerializationException
 import kotools.types.Package
 import kotools.types.SinceKotoolsTypes
+import kotools.types.range.InclusiveBound
+import kotools.types.range.NotEmptyRange
+import kotools.types.range.rangeTo
+import kotools.types.range.toInclusiveBound
 import kotools.types.text.NotBlankString
 import kotools.types.text.toNotBlankString
 import kotlin.jvm.JvmInline
@@ -19,14 +23,36 @@ private constructor(private val value: Int) : NonZeroInt, NegativeInt {
      */
     public companion object {
         /** The minimum value a [StrictlyNegativeInt] can have. */
+        @Deprecated(
+            "Use the range property instead.",
+            ReplaceWith("StrictlyNegativeInt.range.start.value")
+        )
         public val min: StrictlyNegativeInt by lazy(
             Int.MIN_VALUE.toStrictlyNegativeInt()::getOrThrow
         )
 
         /** The maximum value a [StrictlyNegativeInt] can have. */
+        @Deprecated(
+            "Use the range property instead.",
+            ReplaceWith("StrictlyNegativeInt.range.end.value")
+        )
         public val max: StrictlyNegativeInt by lazy(
             (-1).toStrictlyNegativeInt()::getOrThrow
         )
+
+        /** The range of values a [StrictlyNegativeInt] can have. */
+        @SinceKotoolsTypes("4.2")
+        public val range: NotEmptyRange<StrictlyNegativeInt> by lazy {
+            val start: InclusiveBound<StrictlyNegativeInt> = Int.MIN_VALUE
+                .toStrictlyNegativeInt()
+                .getOrThrow()
+                .toInclusiveBound()
+            val end: InclusiveBound<StrictlyNegativeInt> = (-1)
+                .toStrictlyNegativeInt()
+                .getOrThrow()
+                .toInclusiveBound()
+            start..end
+        }
 
         internal infix fun of(value: Int): Result<StrictlyNegativeInt> = value
             .takeIf { it < ZeroInt.toInt() }
@@ -35,7 +61,7 @@ private constructor(private val value: Int) : NonZeroInt, NegativeInt {
 
         /** Returns a random [StrictlyNegativeInt]. */
         @SinceKotoolsTypes("3.0")
-        public fun random(): StrictlyNegativeInt = (min.value..max.value)
+        public fun random(): StrictlyNegativeInt = range.toIntRange()
             .random()
             .toStrictlyNegativeInt()
             .getOrThrow()