+++ src/commonTest/kotlin/kotools/types/range/NotEmptyRangeTest.kt
@@ -8,18 +8,15 @@ import kotlin.test.assertTrue
 class NotEmptyRangeTest {
     @Test
     fun contains_should_pass_with_a_value_in_inclusive_bounds() {
-        val start: InclusiveBound<NonZeroInt> =
-            NonZeroInt.min.toInclusiveBound()
-        val end: InclusiveBound<NonZeroInt> = NonZeroInt.max.toInclusiveBound()
-        val range: NotEmptyRange<NonZeroInt> = start..end
-        val value: NonZeroInt = NonZeroInt.random()
+        val range: NotEmptyRange<PositiveInt> = PositiveInt.range
+        val value: PositiveInt = PositiveInt.random()
         assertTrue { value in range }
     }
 
     @Test
     fun contains_should_pass_with_a_value_in_exclusive_bounds() {
         val start: ExclusiveBound<NonZeroInt> =
-            NonZeroInt.min.toExclusiveBound()
+            NonZeroInt.negativeRange.start.value.toExclusiveBound()
         val end: ExclusiveBound<NonZeroInt> = 2.toStrictlyPositiveInt()
             .getOrThrow()
             .toExclusiveBound()
@@ -32,12 +29,13 @@ class NotEmptyRangeTest {
     @Test
     fun contains_should_pass_with_a_value_in_inclusive_and_exclusive_bounds() {
         val start: InclusiveBound<NonZeroInt> =
-            NonZeroInt.min.toInclusiveBound()
-        val end: ExclusiveBound<NonZeroInt> = NonZeroInt.max.toExclusiveBound()
+            NonZeroInt.negativeRange.start.value.toInclusiveBound()
+        val end: ExclusiveBound<NonZeroInt> =
+            NonZeroInt.positiveRange.end.value.toExclusiveBound()
         val range: NotEmptyRange<NonZeroInt> = start..end
-        val value: NonZeroInt = (NonZeroInt.random()
-            .takeIf { it != NonZeroInt.max }
-            ?: NonZeroInt.min)
+        val value: NonZeroInt = NonZeroInt.random()
+            .takeIf { it != NonZeroInt.positiveRange.end.value }
+            ?: NonZeroInt.negativeRange.start.value
         assertTrue { value in range }
     }
 
@@ -67,10 +65,11 @@ class NotEmptyRangeTest {
     @Test
     fun contains_should_fail_with_a_value_that_is_not_in_inclusive_and_exclusive_bounds() {
         val start: InclusiveBound<NonZeroInt> =
-            NonZeroInt.max.toInclusiveBound()
-        val end: ExclusiveBound<NonZeroInt> = NonZeroInt.min.toExclusiveBound()
+            NonZeroInt.positiveRange.end.value.toInclusiveBound()
+        val end: ExclusiveBound<NonZeroInt> =
+            NonZeroInt.negativeRange.start.value.toExclusiveBound()
         val range: NotEmptyRange<NonZeroInt> = start..end
-        val value: NonZeroInt = NonZeroInt.min
+        val value: NonZeroInt = NonZeroInt.negativeRange.start.value
         assertTrue { value !in range }
     }