+++ changelog.md
@@ -59,6 +59,21 @@ Support for
 [kotlinx.serialization 1.3.3](https://github.com/Kotlin/kotlinx.serialization/releases/tag/v1.3.3)
 (issue [#51](https://github.com/kotools/types/issues/51)).
 
+### Deprecated
+
+The `min` and the `max` companion's properties of the `StrictlyPositiveInt` type
+(issue [#56](https://github.com/kotools/types/issues/56)).
+
+```kotlin
+var result: StrictlyPositiveInt
+// before
+result = StrictlyPositiveInt.min
+result = StrictlyPositiveInt.max
+// after
+result = StrictlyPositiveInt.range.start.value
+result = StrictlyPositiveInt.range.end.value
+```
+
 ## 4.1.0
 
 _Release date: 2023-04-03._