+++ changelog.md
@@ -62,7 +62,7 @@ Support for
 ### Deprecated
 
 The `min` and the `max` companion's properties of the `StrictlyPositiveInt`, the
-`StrictlyNegativeInt` and the `PositiveInt` types (issue
+`StrictlyNegativeInt`, the `PositiveInt` and the `NegativeInt` types (issue
 [#56](https://github.com/kotools/types/issues/56)).
 
 ```kotlin