+++ changelog.md
@@ -71,8 +71,8 @@ implementation("org.kotools:types:$version")
 
 ### Deprecated
 
-The `min` and the `max` companion's properties of the `StrictlyPositiveInt` and
-the `StrictlyNegativeInt` types (issue
+The `min` and the `max` companion's properties of the `StrictlyPositiveInt`, the
+`StrictlyNegativeInt` and the `PositiveInt` types (issue
 [#56](https://github.com/kotools/types/issues/56)).
 
 ```kotlin